Renommer la classe "project" en "project_patch".
