+++ lib/redmine_extended_workflows/controllers/workflows_controller_concern.rb
@@ -13,8 +13,8 @@ module RedmineExtendedWorkflows::Controllers::WorkflowsControllerConcern
       WorkflowProject.replace_permissions(@roles, permissions)
 
       flash[:notice] = l(:notice_successful_update)
-      redirect_to_referer_or if respond_to?(:workflows_permissions_path) # Redmine 4
-      redirect_to_referer_or if respond_to?(:permissions_workflows_path) # Redmine 5
+      redirect_to_referer_or workflows_permissions_path if respond_to?(:workflows_permissions_path) # Redmine < 5
+      redirect_to_referer_or permissions_workflows_path if respond_to?(:permissions_workflows_path) # Redmine >= 5
       return
     end
